requirements.txt
streamlit>=1.32.0
langgraph>=0.1.0
langchain>=0.1.0
langchain-openai>=0.1.0
openai>=1.0.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
youtube-transcript-api>=0.6.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.1.0
python-dotenv>=1.0.0
requests>=2.31.0
lxml>=4.9.0
